Fix dark/light theme title color and refactor
