+++ app/src/main/java/com/github/leuludyha/ibdb/presentation/components/auth/signup/add_friends/AddFriendsHelpers.kt
@@ -5,6 +5,8 @@ import android.content.Context
 import android.content.pm.PackageManager
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import com.github.leuludyha.domain.model.interfaces.Keyed
+import com.github.leuludyha.domain.model.user.User
 
 fun hasContactPermission(context: Context): Boolean {
     // on below line checking if permission is present or not.
@@ -21,4 +23,37 @@ fun requestContactPermission(context: Context, activity: Activity) {
             1
         )
     }
+}
+
+data class UserFromContact(
+    val contact: Contact,
+    val user: User,
+) : Keyed {
+    override fun Id() = user.Id()
+
+}
+
+data class Contact(
+    val name: String,
+    val phoneNumber: String?,
+    val email: String?,
+) {
+
+    private val googleEmailSuffix: String = "@gmail.com"
+
+    /**
+     * @return True if the email is not null and it ends with [googleEmailSuffix]
+     * (It can be used by one of the existing user in the database)
+     */
+    fun isEmailValid() = email?.endsWith(googleEmailSuffix) == true
+
+    /**
+     * @return True if the phone number is not null nor blank, false otherwise
+     */
+    fun isPhoneNumberValid() = !phoneNumber.isNullOrBlank()
+
+    /**
+     * @return True if either the phone number or the email is not null
+     */
+    fun isValid() = isEmailValid() or isPhoneNumberValid()
 }
